decouple loopers
